+++ glyphmaps/Entypo.json
