+++ glyphmaps/SimpleLineIcons.json
