+++ packages/core/src/__tests__/stub.test.ts
