+++ packages/native/src/__tests__/stub.test.ts
