+++ packages/native/src/index.tsx
@@ -1,2 +1,2 @@
-export { default as AudioPlayer } from "./AudioPlayer";
-export { default as Icon } from "./Icon";
+export { default as AudioPlayer } from "./components/AudioPlayer";
+export { default as Icon } from "./components/Icon";
