--- save/mock-backend.json
@@ -1,3 +1,0 @@
-{
-  "components": []
-}
