+++ scripts/get-all-props.js
