+++ scripts/strip-seed-data.js
